+++ packages/dependency-extraction-webpack-plugin/lib/util.js
@@ -1,8 +1,4 @@
 const WORDPRESS_NAMESPACE = '@wordpress/';
-
-// !!
-// This list must be kept in sync with the same list in tools/webpack/packages.js
-// !!
 const BUNDLED_PACKAGES = [
 	'@wordpress/admin-ui',
 	'@wordpress/dataviews',